Add parseAge function to util.js
