+++ changelog.md
@@ -1,15 +1,20 @@
 # Changelog
 
-## [1.0.2]
+## [Unreleased] - v1.0.2-beta.1
+
+### Added
+- Age parsing function to util.js module
 
 ### Changed
 
-- Keep initial PK dosing disclaimer visible after button is clicked to show section.
-- Show link to Bayesian calculator instead of weight-based initial maintenance dose if BMI > 30.
+- Keep initial PK dosing disclaimer visible after button is clicked to show section
+- Show link to Bayesian calculator instead of weight-based initial maintenance dose if BMI > 30 ([#57])
+- Use new parseAge function in form validation for age input
 
 ### Fixed
 
 - Switched from popover to tooltip (age help text wasn't working) ([#56])
+- Allow non-numbers for age input
 
 ## [1.0.1]
 
@@ -152,8 +157,7 @@ Initial POC Release
 ## 0.0.1 - 2021-04-23
 - Initial pre-release for distribution among liaisons/vanco calc workgroup
 
-[Unreleased]: https://github.com/pharmot/multipurpose-calculator/compare/v1.0.2...HEAD
-[1.0.2]: https://github.com/pharmot/multipurpose-calculator/compare/v1.0.1...v1.0.2
+[Unreleased]: https://github.com/pharmot/multipurpose-calculator/compare/v1.0.1...HEAD
 [1.0.1]: https://github.com/pharmot/multipurpose-calculator/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/pharmot/multipurpose-calculator/compare/v0.4.0...v1.0.0
 [0.4.0]: https://github.com/pharmot/multipurpose-calculator/compare/v0.3.0...v0.4.0
@@ -186,3 +190,4 @@ Initial POC Release
 [#43]: https://github.com/pharmot/multipurpose-calculator/issues/43
 [#51]: https://github.com/pharmot/multipurpose-calculator/issues/51
 [#56]: https://github.com/pharmot/multipurpose-calculator/issues/56
+[#57]: https://github.com/pharmot/multipurpose-calculator/issues/57
